loginScreen 함수 분리
