+++ app/src/main/java/com/hansung/sherpa/ui/account/login/LoginScreen.kt
@@ -83,115 +83,6 @@ fun LoginScreen(navController: NavController = rememberNavController(), modifier
     }
 }
 
-@Composable
-fun TitleArea() {
-    Row(verticalAlignment = Alignment.CenterVertically){
-        Icon(
-            imageVector = Icons.Default.Person,
-            contentDescription = "디자인을 위한 사람 아이콘",
-            modifier = Modifier.size(60.dp,60.dp)
-        )
-        Spacer(modifier = Modifier.width(10.dp))
-        Text(
-            text = "로그인",
-            fontSize = 40.sp,
-            fontFamily = BmHanna
-        )
-    }
-}
-
-@Composable
-fun FindAccountArea(navController: NavController) {
-    // 비밀번호 찾기, 회원가입 이동
-    Column(
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        TextButton(
-            onClick = {},
-            colors= ButtonColors(
-                contentColor = Color.Black,
-                containerColor = Color.Transparent,
-                disabledContentColor = Color.Black,
-                disabledContainerColor =  Color.Transparent
-            ),
-            modifier = Modifier.wrapContentWidth()
-        ){
-            Text(
-                text = "아이디/비밀번호를 잃어버리셨나요?",
-                fontSize = 18.sp,
-                fontFamily = BmHanna,
-                style = TextStyle(textDecoration = TextDecoration.Underline)
-            )
-        }
-        TextButton(
-            onClick = {navController.navigate("${SherpaScreen.SignUp.name}")},
-            colors= ButtonColors(
-                contentColor = Color(0xFF34DFD5),
-                containerColor = Color.Transparent,
-                disabledContentColor = Color(0xFF34DFD5),
-                disabledContainerColor =  Color.Transparent
-            ),
-            modifier = Modifier.wrapContentWidth()
-        ){
-            Text(
-                text = "계정이 없으신가요?",
-                fontSize = 18.sp,
-                fontFamily = BmHanna,
-                style = TextStyle(textDecoration = TextDecoration.Underline)
-            )
-        }
-    }
-}
-
-/**
- * 보호자 로그인 구성품
- */
-@Composable
-fun LoginArea(navController: NavController) {
-    val context = LocalContext.current
-
-    var idValue by remember { mutableStateOf("") }
-    var passwordValue by remember { mutableStateOf("") }
-
-    Column(
-        modifier = Modifier
-            .clip(RoundedCornerShape(20.dp))
-            .background(Color.White)
-            .padding(20.dp, 60.dp),
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.spacedBy(30.dp)
-    ) {
-        InfomationGroup("이메일", false) {idValue = it.trim()}
-        InfomationGroup("비밀번호", false) {passwordValue = it.trim()}
-
-        TextButton(
-            onClick = {
-                if(!isValidId(idValue)){
-                    Toast.makeText(context,"로그인 실패!\n아이디 서식을 확인해주세요", Toast.LENGTH_SHORT).show()
-                }
-                if(!isValidId(passwordValue)){
-                    Toast.makeText(context,"로그인 실패!\n비밀번호 서식을 확인해주세요", Toast.LENGTH_SHORT).show()
-                }
-                if(login(navController, idValue, passwordValue)) {
-                    Toast.makeText(context,"로그인 실패!\n아이디 비밀번호를 확인해주세요", Toast.LENGTH_SHORT).show()
-                    return@TextButton
-                } },
-            colors= ButtonColors(
-                contentColor = Color.Black,
-                containerColor = SherpaColor,
-                disabledContentColor = Color.Black,
-                disabledContainerColor =  SherpaColor
-            ),
-            modifier = Modifier.width(200.dp)
-        ){
-            Text(
-                text = "로그인",
-                fontFamily = BmHanna
-            )
-        }
-    }
-}
-
 fun login(navController: NavController, email: String, password: String) : Boolean {
     val loginResponse = UserManager().login(email, password)
